api/t/network: Move IPAM types to their own file
